Rename ImaAdpcmLoader to *Reader and move it to FileFormats
