+++ OpenRA.Mods.Common/OpenRA.Mods.Common.csproj
@@ -147,7 +147,6 @@
     <Compile Include="AI\States\StateBase.cs" />
     <Compile Include="AI\SupportPowerDecision.cs" />
     <Compile Include="AudioLoaders\AudLoader.cs" />
-    <Compile Include="AudioLoaders\ImaAdpcmLoader.cs" />
     <Compile Include="AudioLoaders\VocLoader.cs" />
     <Compile Include="AudioLoaders\WavLoader.cs" />
     <Compile Include="Effects\Beacon.cs" />
@@ -729,6 +728,7 @@
     <Compile Include="Graphics\DetectionCircleRenderable.cs" />
     <Compile Include="Traits\Render\WithIdleAnimation.cs" />
     <Compile Include="FileFormats\AudReader.cs" />
+    <Compile Include="FileFormats\ImaAdpcmReader.cs" />
     <Compile Include="FileFormats\VqaReader.cs" />
     <Compile Include="FileFormats\LCWCompression.cs" />
     <Compile Include="FileFormats\XORDeltaCompression.cs" />
